[SSD222][zkgui] update libeasyui.so to  watermarked version.
[rootcause]:NA
[solution]:NA
[sideeffect]:NA
[ticket]: mantis_12345670
[redflag]: N

+++ bringup/zk_full/jni/include/utils/TimeHelper.h
@@ -18,7 +18,7 @@ public:
 	/**
 	 * @brief 获取当前日期时间，单位为毫秒
 	 */
-	static long getCurrentTime();
+	static long long getCurrentTime();
 
 	/**
 	 * @brief 获取当前日期
